@@ -1,24 +1,14 @@
 ﻿<?xml version="1.0" encoding="utf-8" standalone="yes"?>
 <Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <PublishWithAspNetCoreTargetManifest>false</PublishWithAspNetCoreTargetManifest>
-    <ApplicationInsightsResourceId>/subscriptions/15a2a270-9a75-4f16-bce1-f08edc382c97/resourcegroups/MustakimRG/providers/microsoft.insights/components/JustSending</ApplicationInsightsResourceId>
-    <ApplicationInsightsAnnotationResourceId>/subscriptions/15a2a270-9a75-4f16-bce1-f08edc382c97/resourcegroups/MustakimRG/providers/microsoft.insights/components/JustSending</ApplicationInsightsAnnotationResourceId>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="HangFire" Version="1.7.2" />
-    <PackageReference Include="Hangfire.LiteDB" Version="0.1.1" />
-    <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.6.1" />
+    <PackageReference Include="HangFire" Version="1.7.9" />
+    <PackageReference Include="Hangfire.LiteDB" Version="0.3.1" />
+    <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.12.1" />
     <PackageReference Include="Microsoft.AspNetCore.SignalR" Version="1.1.0" />
     <PackageReference Include="Microsoft.AspNetCore.WebSockets" Version="2.2.1" />
     <PackageReference Include="LiteDB" Version="4.1.4" />
-    <PackageReference Include="Microsoft.AspNetCore.All" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="*.ps1" CopyToPublishDirectory="Always" />
-  </ItemGroup>
-  <ItemGroup>
-    <WCFMetadata Include="Connected Services" />
   </ItemGroup>
 </Project>
