@@ -20,7 +20,7 @@
     <section class="page-header">
         <h1 class="project-name">Secure Line</h1>
         <h2 class="project-tagline">
-            Javascript SDK to enable realtime 1-1 secure message transfer for your user.
+            Javascript SDK to enable realtime 1-1 secure message transfer between your users.
         </h2>
 
         <a href="#doc" class="btn">View documentation</a>
