@@ -1,7 +1,6 @@
-﻿@using Microsoft.ApplicationInsights.AspNetCore
-@inject Microsoft.ApplicationInsights.AspNetCore.JavaScriptSnippet JavaScriptSnippet
+﻿@inject Microsoft.ApplicationInsights.AspNetCore.JavaScriptSnippet JavaScriptSnippet
 <!DOCTYPE html>
-<html>
+<html lang="en">
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0" />
@@ -14,8 +13,8 @@
     <meta property="og:type" content="website">
     <meta property="og:title" content="Just an Email! - Pass files or any texts between your devices">
     <meta property="og:description" content="Quickly send files, URL or any texts across your devices!">
-    <meta property="og:url" content="https://just-an-email.com">
-    <meta property="og:image" content="https://just-an-email.com/images/logo-large.png">
+    <meta property="og:url" content="https://justa.ml">
+    <meta property="og:image" content="https://justa.ml/images/logo-large.png">
     
     <partial name="_CommonHead" />
 
